+++ src/components/SubmitButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import ButtonStyle from '../styles/ButtonStyle';
+import { ButtonStyle } from '../styles';
 
 const Button = styled(ButtonStyle)<{ $gridColumn: string; $gridRow: string }>`
   grid-column: ${props => props.$gridColumn};
@@ -26,11 +26,7 @@ interface SubmitButtonProps {
   gridRow: string;
 }
 
-export default function SubmitButton({
-  text,
-  gridColumn,
-  gridRow,
-}: SubmitButtonProps) {
+export function SubmitButton({ text, gridColumn, gridRow }: SubmitButtonProps) {
   return (
     <Button type="submit" $gridColumn={gridColumn} $gridRow={gridRow}>
       {text}
