+++ src/components/TextButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { styled } from 'styled-components';
 
-import TextButtonStyle from '../styles/TextButtonStyle';
+import { TextButtonStyle } from '../styles';
 
 const Button = styled(TextButtonStyle)<{
   $gridColumn: string;
@@ -24,7 +24,7 @@ interface TextButtonProps {
   onClick: () => void;
 }
 
-export default function TextButton({
+export function TextButton({
   text,
   textSize,
   textColor,
