+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { DatePicker } from './DatePicker';
+export { InformationInput } from './InformationInput';
+export { SubmitButton } from './SubmitButton';
+export { TextButton } from './TextButton';
